- refactor ExtractorsForLong. Getting VARCHAR of '2.5' (or any floating-point) should be allowed  - refactor / improve IfExpression. The rules for determining the return type are  same as those used in CoalesceExpression.  - Removed out-of-date tests in IfExpressionTest  - add test-nullif.yaml

+++ src/main/java/com/akiban/server/types/extract/ExtractorsForLong.java
@@ -56,13 +56,13 @@ class ExtractorsForLong extends LongExtractor {
         case U_DOUBLE:  return (long)source.getUDouble();
         case TEXT:
             try {
-                return Long.parseLong(source.getText());
+                return (long)Double.parseDouble(source.getText());    
             } catch (NumberFormatException ex) {
                 throw new InvalidCharToNumException (source.getText());
             }
         case VARCHAR:
             try {
-                return Long.parseLong(source.getString());
+                return (long)Double.parseDouble(source.getString());               
             } catch (NumberFormatException ex) {
                 throw new InvalidCharToNumException (source.getString());
             }
